+++ iosApp/iosApp/DesignSystem/Components/ErrorBanner.swift
@@ -1,0 +1,18 @@
+import SwiftUI
+
+/// A dismissible error banner for displaying validation or network errors.
+struct ErrorBanner: View {
+    let message: String
+
+    var body: some View {
+        HStack(spacing: 8) {
+            Image(systemName: "exclamationmark.triangle.fill")
+            Text(message)
+        }
+        .font(.subheadline)
+        .foregroundStyle(.red)
+        .padding(12)
+        .frame(maxWidth: .infinity, alignment: .leading)
+        .background(.red.opacity(0.1), in: RoundedRectangle(cornerRadius: 8))
+    }
+}
