+++ iosApp/iosApp/Features/Register/RegisterView.swift
@@ -172,22 +172,6 @@ struct RegisterView: View {
 
 // MARK: - Error Banner (shared component, could be extracted)
 
-private struct ErrorBanner: View {
-    let message: String
-
-    var body: some View {
-        HStack(spacing: 8) {
-            Image(systemName: "exclamationmark.triangle.fill")
-            Text(message)
-        }
-        .font(.subheadline)
-        .foregroundStyle(.red)
-        .padding(12)
-        .frame(maxWidth: .infinity, alignment: .leading)
-        .background(.red.opacity(0.1), in: RoundedRectangle(cornerRadius: 8))
-    }
-}
-
 // MARK: - Previews
 
 #Preview("Register") {
